+++ src/modules/PowerOCR/PowerOCR/App.xaml
@@ -3,8 +3,8 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="clr-namespace:PowerOCR"
-    ShutdownMode="OnExplicitShutdown"
     Exit="Application_Exit"
+    ShutdownMode="OnExplicitShutdown"
     Startup="Application_Startup">
     <Application.Resources />
 </Application>
